requirements.txt
watchdog==2.1.6
pyinstaller==6.3.0
Pillow==10.2.0 